Move IMigrationBuilder out from .Internal
It makes no sense that everyone creating a migration class would have to include the "Internal" namespace. The implementation of the interface, however, makes sense to "hide".
