+++ test/functional/Migration/TestMigration/Migration01.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using RapidCore.Mongo.Migration;
-using RapidCore.Mongo.Migration.Internal;
 
 namespace RapidCore.Mongo.FunctionalTests.Migration.TestMigration
 {
